pyproject.toml
[project]
name = "ec-video-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.0",
    "pyside6>=6.10.2",
]
